In Dockerfile, cut out patches for Pygments and CodeMirror.
A lexer for Maxima has been submitted and accepted to Pygments,
which will presumably appear in the next (2.11 perhaps) release.

A CodeMirror mode for Maxima has been published to npmjs.com
(https://www.npmjs.com/package/codemirror-mode-maxima)
per instructions from CodeMirror, which says new modes should
be published on npmjs instead of being merged into CodeMirror
via pull request.

With this change, colorization is not enabled for the Docker
installation, but in the interest of reducing the complexity
of the installation, I believe it is an acceptable trade-off.